requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
